@@ -20,13 +20,15 @@ final class FiasTableDDLHelper
     /**
      * @throws Exception
      * @throws LogicException
+     * @return list<string>
      */
     public function tableColumnNames(string $tableName): array
     {
         $resultSet = $this->connection->executeQuery(
             sprintf("SELECT array_to_string(tablecolumns('%s'), ',', '*')", $tableName)
         );
-        $result = $resultSet->fetchOne();
+
+        $result = (string)$resultSet->fetchOne();
         if ($result === '') {
             throw new LogicException("No columns found for the table '{$tableName}'");
         }
